+++ source/Messaging/documents/release-notes/release-notes.md
@@ -1,8 +1,9 @@
 # Messaging Release notes
 
-## Version 7.0.1
+## Version 7.1.0
 
-- Update NuGet package dependencies.
+- Mark dependency injection extensions `ServiceBusExtensions.AddServiceBusClientForApplication` as obsolete, and implemented an overloaded version that accepts a `tokenCredentialFactory`.
+- Mark dependency injection extensions `ServiceBusExtensions.AddDeadLetterHandlerForIsolatedWorker` as obsolete, and implemented an overloaded version that accepts a `tokenCredentialFactory`.
 
 ## Version 7.0.0